Modern NEWS WEBSITE
This above code is of responsive News website.
By using HTML, CSS and JavaScript this news website can be prepared. In this website there are three li and one search bar which is used to search the news in a navbar.
when we click on logo of website, it will take us on Home of the website.